watcher/requirements-web-console.txt
pystray
pillow

# Optional: package into .exe later
# pyinstaller
